Corrected Delta Time calculation
float was not enough to calculate an accurate delta time between frames, the physics step would not appear smooth. floats were truncating the tick delta causing the delta to not adjust as the time between varies. x64 issues...

+++ include/core/ChunkyVoxelsMain.h
@@ -36,7 +36,7 @@ public:
     void operator=(ChunkyVoxelsMain const&) = delete;
 
     void initialize(HWND hWnd, HINSTANCE hInstance);
-    void tick(float fDt);
+    void tick(double fDt);
     void shutdown(void);
